Unsigned compared against zero with CID 147540-42
issues:
coverity scan CID:147540, function: zpios_timespec_normalize
coverity scan CID:147542, function: spa_prop_validate

+++ include/zpios-ctl.h
@@ -146,7 +146,7 @@ zpios_timespec_normalize(zpios_timespec_t *ts, uint32_t sec, uint32_t nsec)
 		nsec -= NSEC_PER_SEC;
 		sec++;
 	}
-	while (nsec < 0) {
+	while (((int32_t)nsec) < 0) {
 		nsec += NSEC_PER_SEC;
 		sec--;
 	}
